refactor: typo fixtures
